Moved schemaDialect() from driver dialect traits to drivers
